+++ modules/blogging/src/Volo.Blogging.Admin.Application/Volo/Blogging/Admin/BloggingAdminApplicationAutoMapperProfile.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using Volo.Blogging.Admin.Blogs;
 using Volo.Blogging.Blogs;
+using Volo.Blogging.Blogs.Dtos;
 
 namespace Volo.Blogging.Admin
 {
