+++ modules/blogging/src/Volo.Blogging.Admin.Web/AbpBloggingAdminWebAutoMapperProfile.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
 using Volo.Blogging.Admin.Blogs;
 using Volo.Blogging.Admin.Pages.Blogging.Admin.Blogs;
+using Volo.Blogging.Blogs;
+using Volo.Blogging.Blogs.Dtos;
 using EditModel = Volo.Blogging.Admin.Pages.Blogging.Admin.Blogs.EditModel;
 
 namespace Volo.Blogging.Admin
