--- modules/blogging/src/Volo.Blogging.Application.Contracts/Volo/Blogging/Blogs/Dtos/BlogDto.cs
@@ -1,14 +1,0 @@
-using System;
-using Volo.Abp.Application.Dtos;
-
-namespace Volo.Blogging.Blogs.Dtos
-{
-    public class BlogDto : FullAuditedEntityDto<Guid>
-    {
-        public string Name { get; set; }
-
-        public string ShortName { get; set; }
-
-        public string Description { get; set; }
-    }
-}
